Code refactor bettween child_sync and sponsorship_sync
- Don't put contract in Biennial state if last Biennial in GP is more
  recent than the last Case Study in Odoo.
