+++ sponsorship_sync_gp/model/__init__.py
@@ -10,5 +10,6 @@
 ##############################################################################
 
 from . import gp_connector
+from . import child
 from . import contracts
 from . import account_invoice
